+++ sources/Google.Solutions.Mvvm/Google.Solutions.Mvvm.csproj
@@ -105,7 +105,6 @@
     <Compile Include="Drawing\HslColor.cs" />
     <Compile Include="Drawing\IconInverter.cs" />
     <Compile Include="Format\MarkdownDocument.cs" />
-    <Compile Include="Format\MarkdownRtfConverter.cs" />
     <Compile Include="Format\RtfWriter.cs" />
     <Compile Include="Interop\SubclassCallback.cs" />
     <Compile Include="Interop\WindowMessage.cs" />
